EID-1666 Handle eIDAS unsigned assertions
The eIDAS SAML spec doesn't require countries to sign their assertions.
Some do, some don't. There has been extensive work done in the hub to
handle this whilst maintaining end to end trust for assertions sent to
the MSA and the VSP. See various PR's on those repo's for more detail.

This PR implements handling eIDAS SAML responses being sent to the VSP
as an assertion attribute in a wrapper Response from the hub, and
decrypting the assertions contained within using encrypted keys, also
sent as assertion attribute.

The most significant changes here are the logic changes in the
`ResponseService` to use a new `UnsignedAssertionsResponseHandler` to
extract the eIDAS assertions contained in the wrapped SAML Response.

Also, changes to the `EidasAssertionTranslator` to allow for a new
`EidasUnsignedAssertionTranslator` to be created without repeating
logic. The only difference between the two is the lack of validating the
assertions signature in the latter.

There is a new class, `EidasEncryptionAlgorithmValidator`. The eIDAS
spec insists on the use of certain encryption algorithms which differ
from the hub. This is where they're configured.