EID-1607 Throw if no unsigned assertions handler in eIDAS flow
The ResponseService only needs to be configured with an
UnsignedAssertionsResponseHandler if we're going to be dealing with
eIDAS response. This removes the Optional and throws if the handler
isn't present when required. Currently it should be impossible to get
into this situation, but it will help guard against this happening
accidentally in future changes to the VSP.